feat: User info and tidying
