+++ examples/nextjs-comments-canvas/src/components/Toolbar.module.css
@@ -52,3 +52,21 @@
   border-radius: 8px;
   cursor: pointer;
 }
+
+.newThreadCursor {
+  position: fixed;
+  top: 0;
+  left: 0;
+  pointer-events: none;
+  user-select: none;
+  width: 36px;
+  height: 36px;
+  box-shadow: 0 0 0 1px rgb(0 0 0 / 4%), 0 2px 6px rgb(0 0 0 / 4%), 0 8px 26px rgb(0 0 0 / 6%);;
+  border-top-left-radius: 4px;
+  border-top-right-radius: 50%;
+  border-bottom-right-radius: 50%;
+  border-bottom-left-radius: 50%;
+  cursor: grab;
+  background: #fff;
+}
+
